W06-W07/Projet/Transfer-Learning/requirements.txt
ipywidgets
ipykernel
kagglehub
pillow
pandas
matplotlib
opencv-python
numpy
tensorflow
mlflow
python-dotenv